remodularização dos slots
